Update hostapd config file handling
For JB MR1 we do not want to install default versions of hostapd data
files in either the data or persist partitions.  Instead all default
data files should first start in the system partition, and the
appropriate runtime files will be created on the data partition by the
SoftAP SDK framework as needed.

Change-Id: I8bc7dfc89f5c1ec574ffdb54afdee2e2d28196e7