wlan: update logic to populate WCNSS_qcom_wlan_nv.bin
In MR1-based releases we have expected the nv.bin file to be present
on the data partition.  If it did not exist, it would be created,
either from the version on the persist partition (if it had been
created there by the manufacturing process) or from a "canned" version
from the system partition, created there by the build process.  In
particular we did not update the nv.bin on the data partition if the
file already existed.  This meant that if a device had an nv.bin on
the data partition, and if it's nv.bin on the persist partition was
updated by the manufacturing process, the updated nv.bin on persist
was not being used.

To avoid this issue update the logic to conform to the pre-MR1
behavior.  There will only be one nv.bin on the device, and it will
exist on the persist partition.

Change-Id: I191daa3661666f6369745ab22b2726485707bb36
CRs-fixed: 448163