Move set of rmnet interfaces later in boot process
telephony: move set of accept_ra==2 for rmmnet interfaces
later in the boot process to ensure that the interfaces
are available for writing.

background: Set accept_ra==2 for rmnet interfaces

When ipv6 forwarding is active, rmnet interfaces
are not receiving an ipv6 address. This will happen for
instance, when tethering is enabled before the rmnet
interface is setup.

Set accep_ra==2, for those interfaces, which allows them
to act on router advertisements even when forwarding
is active

Change-Id: If0e4468ef300135c6de8a6b159fc3f2810edb0ed