wlan: updating opensource driver config files
wlan: Add config param AP LINK MONITOR TIMEOUT
Add AP LINK MONITOR TIMEOUT to change the frequency of KEEP ALIVE
packets sent during the KEEP ALIVE TIMEOUT duration for SAP and GO.
It is used to change the frequency of keep alive packets in the AP
Link Monitor period which is by default 20s. The keep alive packets
are sent as an interval of 3s which can be changed by setting
gApLinkMonitorPeriod in ini file.

Change-Id: Ie1544031f9a8546562ca924e8921a60f0b0327b7
CRs-fixed: 425674